Image Processing/requirements.txt
astroid==2.3.1
beautifulsoup4==4.8.0
bs4==0.0.1
cpbd==1.0.7
cycler==0.10.0
decorator==4.4.0
ExifRead==2.1.2
hurry==1.1
hurry.filesize==0.9
imageio==2.5.0
imutils==0.5.3
isort==4.3.21
joblib==0.13.2
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
matplotlib==3.1.1
mccabe==0.6.1
networkx==2.3
np==1.0.2
numpy==1.17.2
opencv-python==4.1.1.26
pandas==0.25.1
piexif==1.1.3
Pillow==6.1.0
pylint==2.4.2
pyparsing==2.4.2
pytesseract==0.3.0
python-dateutil==2.8.0
pytz==2019.2
PyWavelets==1.0.3
pyzbar==0.1.8
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.1.0
sewar==0.4.2
six==1.12.0
sklearn==0.0
soupsieve==1.9.3
typed-ast==1.4.0
wrapt==1.11.2
